This commit was manufactured by cvs2svn to create tag 'v0-12-2'.

+++ PlugIns/GuiElements/include/OgrePanelGuiElement.h
@@ -110,7 +110,7 @@ namespace Ogre {
         // Texture tiling
         Real mTileX[OGRE_MAX_TEXTURE_LAYERS];
         Real mTileY[OGRE_MAX_TEXTURE_LAYERS];
-        size_t mNumTexCoordsInBuffer;
+        unsigned short mNumTexCoordsInBuffer;
 
         RenderOperation mRenderOp;
